Added login/logout buttons in Poll and Results. Added logging for submission. Closes: #9 Closes: #12
